Fix issue around marketviewmodel caching
This code around Market handling and specifically the use of CacheManager needs to be reviewed. An example is the AddMarketViewModel method. It is executed for every visitor to the site, on every request. It uses the CacheManager which is global for the entire application. This means it is not possible to have more than one user in different markets without strange results.

To fix cache the MarketViewModel taking the user's market into account so other users in the same market can take advantage of the cached model.

To recreate:

- Open a browser and browse the site using the USA market
- Open a incognito window and browse to site, select the UK market
- Back in the non-incognito market refresh
- In the incognito window refresh
- Note that the incognito window is show UK as the market but USA as prices